Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
python-barcode qrcode[pil]
python-barcode pillow pyzbar